Model i2c_core more accurately
`i2c_state` have additional 9 fields to represent signals and registers
in `i2c_core`.  Most notably is the `fsm_state` which mimics the states
in `i2c_fsm`.

FSM States
----------

Some of the states in `i2c_fsm` are grouped together for modularity
reason. For example, states such as `ClockLow`, `ClockPulse`, `HoldBit`,
`ClockLowAck`, `ClockPulseAck`, `HoldDevAck` are grouped together under
the constructor `Transmitting`.

Some of the states in `i2c_fsm` are abstracted away. For example,
`HostClockLowAck` and `HostClockPulseAck` are simply modelled as
`HostClock` here.

Delays
------

Many states in `i2c_fsm` are assigned with delays, which are set up via
1.) TIMING0, TIMING1, TIMING2, TIMING3 registers, and 2.) some
mathematical formula described in the documentation (e.g. some requires
the addition of rising and falling times).  This is modelled via the
variable `counter'`.

RX_FIFO
-------

The non-determinism to decide when to insert a byte to `rx_fifo` has
now been replaced with a guard about `fsm_state`. Only the data is
modelled via non-determinism `fnums`.

Registers
---------

The `i2c_tick` now has a field for registers, modelling the interrupts
`cmd_complete` and `nak`.

Signed-off-by: Albert Rizaldi <albert.rizaldi@planv.tech>