Model clock stretching for the counter
The model of the counter was not faithful because it did not model the
clock stretching. This commit improves the model by adding clock
stretching.

To do so, all FSM states needs to be exactly the same as in the RTL
code.

This commit also includes some cosmetic change (mainly indentation).

Signed-off-by: Albert Rizaldi <albert.rizaldi@planv.tech>