Make `tick` fallible
The inciting reason to do this was to be able to model window bus
errors; it could also be useful for disallowing certain actions, though
(e.g. making sure the I2C peripheral is never set to target mode).

Whether or not the latter use case is supported comes down to whether
the `rsp.error <=> ISL (cheshire_run ...)` clause is preserved, or if
it's changed to `rsp.error ==> ISL (cheshire_run ...)`.