Fix spi_hostRegsCircuitProofTheory with window support
There were a lot of issues that i2c didn't surface due to not using
windows...

Something else notable here is that we're assuming windows always
respond immediately. This isn't true in general, but would require a lot
of messing with the register framework to fix when I doubt we're going
to encounter a Cheshire peripheral this isn't true for. When we start
doing non-Cheshire peripherals, we shouldn't make this assumption there.